+++ Packages/com.unity.inputsystem/InputSystem/Editor/UITKAssetEditor/Commands/Commands.cs
@@ -97,6 +97,90 @@ namespace UnityEngine.InputSystem.Editor
             };
         }
 
+        public static Command CopyActionMapSelection()
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                CopyPasteHelper.CopyActionMap(state);
+                return state;
+            };
+        }
+
+        public static Command CutActionMapSelection()
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                CopyPasteHelper.CopyActionMap(state);
+                return DeleteActionMap(state.selectedActionMapIndex).Invoke(state);
+            };
+        }
+
+        public static Command CopyActionBindingSelection()
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                CopyPasteHelper.Copy(state);
+                return state;
+            };
+        }
+
+        public static Command CutActionsOrBindings()
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                CopyPasteHelper.Copy(state);
+                return state.selectionType == SelectionType.Action ?
+                    DeleteAction(state.selectedActionMapIndex, Selectors.GetSelectedAction(state)?.wrappedProperty.FindPropertyRelative(nameof(InputAction.m_Name)).stringValue).Invoke(state)
+                    : DeleteBinding(state.selectedActionMapIndex, state.selectedBindingIndex).Invoke(state);
+            };
+        }
+
+        public static Command PasteActionMaps()
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                var lastPastedElement = CopyPasteHelper.PasteActionMapsFromClipboard(state);
+                if (lastPastedElement != null)
+                {
+                    state.serializedObject.ApplyModifiedProperties();
+                    return state.SelectActionMap(lastPastedElement.GetIndexOfArrayElement());
+                }
+                return state;
+            };
+        }
+
+        public static Command PasteActionFromActionMap()
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                var lastPastedElement = CopyPasteHelper.PasteActionsOrBindingsFromClipboard(state, true);
+                if (lastPastedElement != null)
+                {
+                    state.serializedObject.ApplyModifiedProperties();
+                    return state.SelectAction(lastPastedElement.GetIndexOfArrayElement());
+                }
+                return state;
+            };
+        }
+
+        public static Command PasteActionsOrBindings()
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                var typeOfCopiedData = CopyPasteHelper.GetCopiedClipboardType();
+                var lastPastedElement = CopyPasteHelper.PasteActionsOrBindingsFromClipboard(state);
+                if (lastPastedElement != null)
+                {
+                    state.serializedObject.ApplyModifiedProperties();
+                    if (typeOfCopiedData == typeof(InputAction))
+                        return state.SelectAction(lastPastedElement.GetIndexOfArrayElement());
+                    if (typeOfCopiedData == typeof(InputBinding))
+                        return state.SelectBinding(lastPastedElement.GetIndexOfArrayElement());
+                }
+                return state;
+            };
+        }
+
         public static Command DuplicateActionMap(int actionMapIndex)
         {
             return (in InputActionsEditorState state) =>
@@ -104,7 +188,7 @@ namespace UnityEngine.InputSystem.Editor
                 var actionMapArray = state.serializedObject.FindProperty(nameof(InputActionAsset.m_ActionMaps));
                 var actionMap = Selectors.GetActionMapAtIndex(state, actionMapIndex)?.wrappedProperty;
                 var name = actionMap?.FindPropertyRelative(nameof(InputAction.m_Name)).stringValue;
-                var newMap = InputActionSerializationHelpers.DuplicateElement(actionMapArray, actionMap, name, actionMap.GetIndexOfArrayElement() + 1);
+                var newMap = CopyPasteHelper.DuplicateElement(actionMapArray, actionMap, name, actionMap.GetIndexOfArrayElement() + 1);
                 state.serializedObject.ApplyModifiedProperties();
                 return state.SelectActionMap(newMap.FindPropertyRelative(nameof(InputAction.m_Name)).stringValue);
             };
@@ -115,10 +199,9 @@ namespace UnityEngine.InputSystem.Editor
             return (in InputActionsEditorState state) =>
             {
                 var action = Selectors.GetSelectedAction(state)?.wrappedProperty;
-                var actionName = action?.FindPropertyRelative(nameof(InputAction.m_Name)).stringValue;
                 var actionMap = Selectors.GetActionMapAtIndex(state, state.selectedActionMapIndex)?.wrappedProperty;
                 var actionArray = actionMap?.FindPropertyRelative(nameof(InputActionMap.m_Actions));
-                InputActionSerializationHelpers.DuplicateAction(actionMap, actionArray, action, actionName);
+                CopyPasteHelper.DuplicateAction(actionArray, action, actionMap, state);
                 state.serializedObject.ApplyModifiedProperties();
                 return state.SelectAction(state.selectedActionIndex + 1);
             };
@@ -132,7 +215,7 @@ namespace UnityEngine.InputSystem.Editor
                 var actionName = binding?.FindPropertyRelative("m_Action").stringValue;
                 var actionMap = Selectors.GetActionMapAtIndex(state, state.selectedActionMapIndex)?.wrappedProperty;
                 var bindingsArray = actionMap?.FindPropertyRelative(nameof(InputActionMap.m_Bindings));
-                var newIndex = InputActionSerializationHelpers.DuplicateBinding(bindingsArray, binding, actionName, binding.GetIndexOfArrayElement() + 1);
+                var newIndex = CopyPasteHelper.DuplicateBinding(bindingsArray, binding, actionName, binding.GetIndexOfArrayElement() + 1);
                 state.serializedObject.ApplyModifiedProperties();
                 return state.SelectBinding(newIndex);
             };
@@ -141,7 +224,7 @@ namespace UnityEngine.InputSystem.Editor
         private static InputActionsEditorState SelectPrevActionMap(InputActionsEditorState state)
         {
             var count = Selectors.GetActionMapCount(state);
-            int index = 0;
+            var index = 0;
             if (count != null && count.Value > 0)
                 index = Math.Max(state.selectedActionMapIndex - 1, 0);
             return state.SelectActionMap(index);
@@ -153,7 +236,6 @@ namespace UnityEngine.InputSystem.Editor
             {
                 var actionMap = Selectors.GetActionMapAtIndex(state, actionMapIndex)?.wrappedProperty;
                 var action = Selectors.GetActionInMap(state, actionMapIndex, actionName).wrappedProperty;
-                var actionIndex = action.GetIndexOfArrayElement();
                 var actionID = InputActionSerializationHelpers.GetId(action);
                 InputActionSerializationHelpers.DeleteActionAndBindings(actionMap, actionID);
                 state.serializedObject.ApplyModifiedProperties();
@@ -175,16 +257,6 @@ namespace UnityEngine.InputSystem.Editor
                 // ActionsTreeView will dispatch a separate command to select the previous Binding
                 return state;
             };
-        }
-
-        public static Command ExpandCompositeBinding(SerializedInputBinding binding)
-        {
-            return (in InputActionsEditorState state) => state.ExpandCompositeBinding(binding);
-        }
-
-        public static Command CollapseCompositeBinding(SerializedInputBinding binding)
-        {
-            return (in InputActionsEditorState state) => state.CollapseCompositeBinding(binding);
         }
 
         public static Command SelectBinding(int bindingIndex)
@@ -227,7 +299,7 @@ namespace UnityEngine.InputSystem.Editor
         {
             return (in InputActionsEditorState state) =>
             {
-                InputActionSerializationHelpers.ChangeBinding(bindingProperty.wrappedProperty, partName);
+                InputActionSerializationHelpers.SetBindingPartName(bindingProperty.wrappedProperty, partName);
                 state.serializedObject.ApplyModifiedProperties();
                 return state;
             };